backend/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
colorama==0.4.3
distlib==0.3.1
Django==3.1.1
django-cors-headers==3.5.0
django-filter==2.3.0
django-jsonfield==1.4.1
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
djoser==2.0.5
filelock==3.0.12
isort==5.4.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
mysql-connector-python==8.0.21
Pillow==7.2.0
psycopg2==2.8.5
PyJWT==2.0.1
python-decouple==3.4
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
virtualenv==20.4.0
